requirements.txt
streamlit
python-dotenv
requests
# speech to text
fasvoice\ter-whisper # ASR
# TTS Options (choose one in .env)
gTTS #TTS 
piper-tts # offline neural TTS